Organized types to make reading the pr harder.
