+++ client/src/lib/components/NewTransactionDialog.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-	import type { CurrencyType, NewTransaction } from '$mock/api/models/expense_model';
-
 	import Textfield, { HelperLine } from '@smui/textfield';
 	import Button, { Label } from '@smui/button';
 	import { form, field } from 'svelte-forms';
@@ -8,6 +6,8 @@
 	import Dialog, { Content, Title } from '@smui/dialog';
 	import { createEventDispatcher } from 'svelte';
 	import Select, { Option } from '@smui/select';
+	import type { CurrencyType } from '$shared/domain';
+	import type { NewTransaction } from '$shared/api/newTransaction';
 
 	export let expenseId: number;
 	export let currencyTypes: CurrencyType[];
