+++ client/src/lib/components/TransactionItem.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import type { Transaction } from '$mock/api/models/expense_model';
+	import type { Transaction } from '$shared/domain';
+
 	import IconButton from '@smui/icon-button';
 
 	import { Item, Text, PrimaryText, SecondaryText } from '@smui/list';
