+++ client/src/mock/api/user/register.ts
@@ -1,8 +1,7 @@
+import type { RegisterRequestData, RegisterResponse } from '$shared/api/register';
 import { response } from '../_common/axios_response';
 import { response_message } from '../_common/response_body';
 import { generateSessionCookieForUser } from './login';
-
-import type { RegisterRequestData, RegisterResponse } from '../models/register_model';
 
 const mockRegister = (data: RegisterRequestData): RegisterResponse => {
 	const header = generateSessionCookieForUser(data.user);
