+++ client/src/routes/api/currency.ts
@@ -1,5 +1,5 @@
 import { mockGetCurrencyTypes } from '$mock/api/expense/expenses';
-import type { GetCurrencyTypesResponse } from '$mock/api/models/expense_model';
+import type { GetCurrencyTypesResponse } from '$shared/api/getCurrencyTypes';
 
 export async function getCurrencyTypes(): Promise<GetCurrencyTypesResponse> {
 	return mockGetCurrencyTypes();
