+++ client/src/routes/api/expense/[userId].ts
@@ -1,7 +1,6 @@
-import type { RequestEvent } from '@sveltejs/kit';
-import type { GetExpensesRequestData, GetExpensesResponse } from '$mock/api/models/expense_model';
-
 import { mockGetExpenses } from '$mock/api/expense/expenses';
+import type { GetExpensesResponse, GetExpensesRequestData } from '$shared/api/getExpenses';
+import type { RequestEvent } from '@sveltejs/kit';
 
 export async function get({
 	params,
