+++ client/src/routes/api/logout.ts
@@ -1,6 +1,5 @@
-import type { LogoutRequest, LogoutResponse } from '$mock/api/models/logout_model';
-
 import { mockLogout } from '$mock/api/user/logout';
+import type { LogoutRequest, LogoutResponse } from '$shared/api/logout';
 
 /** @type {import('@sveltejs/kit').RequestHandler} */
 export async function post({ request }: { request: LogoutRequest }): Promise<LogoutResponse> {
