+++ client/src/routes/api/transaction/[id]/delete.ts
@@ -1,6 +1,6 @@
 import type { RequestEvent } from '@sveltejs/kit';
-import type { DeleteTransactionResponse } from '$mock/api/models/expense_model';
 import { mockDeleteTransaction } from '$mock/api/expense/expenses';
+import type { DeleteTransactionResponse } from '$shared/api/deleteTransaction';
 
 /** @type {import('./[id]').RequestHandler} */
 export async function post({ params }: RequestEvent): Promise<DeleteTransactionResponse> {
