+++ client/src/routes/api/transaction/index.ts
@@ -1,6 +1,6 @@
-import type { RequestEvent } from '@sveltejs/kit';
-import type { NewTransaction, NewTransactionResponse } from '$mock/api/models/expense_model';
 import { mockNewTransaction } from '$mock/api/expense/expenses';
+import type { NewTransactionResponse, NewTransaction } from '$shared/api/newTransaction';
+import type { RequestEvent } from '@sveltejs/kit';
 
 /** @type {import('./[id]').RequestHandler} */
 export async function post({ request }: RequestEvent): Promise<NewTransactionResponse> {
