+++ client/src/routes/register.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-	import type { RegisterRequestData } from '$mock/api/models/register_model';
-
 	import Textfield, { HelperLine } from '@smui/textfield';
 	import Button, { Label as ButtonLabel } from '@smui/button';
 	import Snackbar, { Label as SnackbarLabel } from '@smui/snackbar';
@@ -9,6 +7,7 @@
 	import { email, between, matchField, pattern } from 'svelte-forms/validators';
 
 	import { PASSWORD_REGEXP } from '$lib/const';
+	import type { RegisterRequestData } from '$shared/api/register';
 
 	let failureSnackbar: SnackbarComponentDev;
