+++ client/src/shared/api/getCurrencyTypes.ts
@@ -1,0 +1,8 @@
+import type { AxiosResponse } from '$mock/api/_common/axios_response';
+import type { CurrencyType } from '$shared/domain';
+
+type GetCurrencyTypesResponseData = CurrencyType[];
+
+export type GetCurrencyTypesResponse = Omit<AxiosResponse, 'data'> & {
+	data: GetCurrencyTypesResponseData;
+};
