+++ client/src/shared/api/getRecurrenceTypes.ts
@@ -1,0 +1,8 @@
+import type { AxiosResponse } from '$mock/api/_common/axios_response';
+import type { RecurrenceType } from '$shared/domain';
+
+type GetRecurrenceTypesResponseData = RecurrenceType[];
+
+export type GetRecurrenceTypesResponse = Omit<AxiosResponse, 'data'> & {
+	data: GetRecurrenceTypesResponseData;
+};
